+++ data/core.lua
@@ -1,4 +1,6 @@
--- Final Fantasy Level Up! v2.0.1
+--=====================================================================================
+-- FFLU | Final Fantasy Level-Up! - core.lua
+--=====================================================================================
 
 local FFLU = {} -- Create a namespace to avoid global variable pollution
 
@@ -12,9 +14,8 @@ function FFLU.MuteDefaultLevelUpSound()
     MuteSoundFile(569593) -- Mutes the default level up sound
 end
 
-
--- Display a message on player login
-local function DisplayWelcomeMessage()
+-- Display a welcome message on player login
+function FFLU.DisplayWelcomeMessage()
     print("[|Tinterface/addons/FFLU/images/icon:16:16|t - |cffffe568F|r|cffffffffinal|r |cffffe568F|r|cffffffffantasy|r |cffffe568L|r|cffffffffevel|r |cffffe568U|r|cffffffffp!|r] |cffffe568Loaded!|r |cff8080ff(v2.1.3)|r")
 end
 
@@ -27,7 +28,7 @@ frame:RegisterEvent("PLAYER_LOGIN")
 frame:SetScript("OnEvent", function(self, event, ...)
     if event == "PLAYER_LEVEL_UP" then
         FFLU.PlayCustomLevelUpSound()
-    elseif event == "ADDON_LOADED" and ... == "FFLU" then -- Check addon name
+    elseif event == "ADDON_LOADED" and ... == "FFLU" then
         FFLU.MuteDefaultLevelUpSound()
     elseif event == "PLAYER_LOGIN" then
         FFLU.DisplayWelcomeMessage()